# packages/backend/openapi.yaml
openapi: 3.1.0
info:
  title: BRACE Backend
  version: 0.2.0
paths:
...
  /api/size/calc:
    post:
      tags:
      - Main Screen
      summary: Calculate Size
      operationId: calculate_size_api_size_calc_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SizeCalculationRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse_SizeCalculationResponse_'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
components:
  schemas:
...
    ErrorResponse:
      properties:
        type:
          type: string
          title: Type
        message:
          type: string
          title: Message
      type: object
      required:
      - type
      - message
      title: ErrorResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    Pagination:
      properties:
        page:
          type: integer
          minimum: 1.0
          title: Page
        page_size:
          type: integer
          maximum: 200.0
          minimum: 1.0
          title: Page Size
        total:
          type: integer
          minimum: 0.0
          title: Total
        pages:
          type: integer
          minimum: 1.0
          title: Pages
      type: object
      required:
      - page
      - page_size
      - total
      - pages
      title: Pagination
...
    SizeCalculationRequest:
      properties:
        waist:
          type: number
          exclusiveMinimum: 0.0
          title: Waist
        hip:
          type: number
          exclusiveMinimum: 0.0
          title: Hip
      type: object
      required:
      - waist
      - hip
      title: SizeCalculationRequest
    SizeCalculationResponse:
      properties:
        size:
          type: string
          title: Size
      type: object
      required:
      - size
      title: SizeCalculationResponse
...
    SuccessResponse_SizeCalculationResponse_:
      properties:
        data:
          $ref: '#/components/schemas/SizeCalculationResponse'
        error:
          anyOf:
          - $ref: '#/components/schemas/ErrorResponse'
          - type: 'null'
        pagination:
          anyOf:
          - $ref: '#/components/schemas/Pagination'
          - type: 'null'
      type: object
      required:
      - data
      title: SuccessResponse[SizeCalculationResponse]
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError